backend/requirements.txt
django==5.0.3
djangorestframework==3.15.1
django-cors-headers==4.3.1
qrcode[pil]>=8.0
pillow>=9.1.0
google-genai>=1.0.0
fpdf2>=2.8.0
pdfplumber>=0.11.0
python-docx>=1.1.0
python-dotenv>=1.0.0
